Optimize some parameters.
Now the memory occupied by tx_data_manager is not counted by cache
manager, so we should decrease the ledger_db_size to keep the total
memory usage at almost the same level.

The current MAX_DIRTY_AND_TEMPORARY_TRIE_NODES may not be sufficient if we
have both high throughput and large number of accounts.

Timeout is spotted for headers/blocks during catch-up, so increase it to avoid
timeout for honest peers.