Bug fix in #1182. (#1228)
The bug is caused in #1182 when I merged two cases for CompressedPathRaw creation, however new_and_apply_mask assumes that the length of the compressed path is at least one, thus the bug.

I didn't modify the changelog because it's the bug introduced and fixed within the same version.

The bug triggers only in debug test mode. I'm thinking about running debug tests periodically as well.