Only use small expire_timeout for tests that involve both era and crash.
`reorg_test` may fail before because when blocks are expired, not new
blocks are generated to trigger requesting them again. And it's possible
to expire blocks in `reorg_test` because it will request the blocks in
the fork chain of another shard one by one.

We need expire_gc to be triggered to mark blocks before era_genesis as
graph-ready, so it's only needed if both checkpoints and crash will
happen.