Make estimateGasAndCollateral work for "reverted" transaction (#1273)
Reverted is a special vm error code. For a reverted transaction we will charge only for used gas, therefore estimateGas should not give an error.

In call_virtual and estimateGasAndCollateral, when using a random sender address, make sure the sender address is a user address. And use AddressUtil wherever it applies.