Fix wrong epoch hash in db for receipts.
In the old implementation, if

1. Block A is on pivot chain and executed.
2. Pivot chain switches, and Block A is referred and executed in the epoch of block B.
3. Pivot chain switches back where A is on pivot chain.

Our receipts for Block A in db will still be the one that it's executed in
the epoch of block B.

This bugfix is just a patch, we should refactor/optimize this receipt
storage to make it more clear and efficient. The current implemenration may
need to insert TransactionIndex and BlockExecutionResult duplicatedly
due to pivot chain switch, especially when optimistic execution is
enabled.