Disable tx index recording by default.
Recording tx index is only needed to serve tx-related RPCs, and will not
be needed for miners. The tx index takes about 40% of `blockchain_db`
for full nodes and it will not be removed in making checkpoints, so it's
better to be disabled by default. Those RPC providers should enable it
manually.