Fix some random test failure. (#1663)
There full node tests may fail as a part of rigorous testing when run on
a relatively slow Azure VM.

For `sync_checkpoint_test.py`, it fails if we query states before the
node finishes processing all blocks.

For `remove_old_eras_test.py`, it fails if there are some forks in the
generated blocks. Note that we update best_info after releasing
consensus inner lock, so it's possible that a generated block is not
seen by the next call of `generate_one_block`.