Use upgradable ChainIdParams (#2000)
The ChainIdParams is designed to return a chain_id for specified block height. The design is useful to describe hard-forks. This PR completes the implementation of the design, and makes it possible to specify chain_id changes by block heights in configuration.

As part of this PR, there is a static upgradable CHAIN_ID params shared by all components. At this moment, MOST of these components DO NOT EXPECT the chain_id to change.

It's worth noting that, if chain_id and chain_id_params is unspecified in config, which is often the case for intranet test deployment, the chain_id will be randomly generated, and the network_id will be the same as the randomly generated chain_id. In this case, the old behavior was to set the network_id 1. The change of behavior is negligible.

There is no change to existing network messages. There is no known plan to start a hard-fork yet.