Increase request delay upper bound to tolerate corner cases.
If a block is not received after this upper bound, the request will be
removed, and this will prevent all its futures from becoming graph-ready.
If some block is just sent before the final timeout boundary, it may
cause a temporary fork because some nodes will not receive it.

Now it's increased so that benign stale blocks are not likely to cause
this issue.